LayoutTests/fast/text/international/system-language/navigator-language/navigator-language-ru-2.html

<!DOCTYPE html><!-- webkit-test-runner [ language=ru ] -->
<html>
<head>
<script src="../../../../../resources/js-test.js"></script>
</head>
<body>
<script>
shouldBeEqualToString("navigator.language", "ru");
</script>
</body>
</html>
