LayoutTests/fast/text/international/system-language/navigator-language/navigator-language-ru-2-expected.txt

PASS navigator.language is "ru"
PASS successfullyParsed is true

TEST COMPLETE
